% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_hmm_transitions}
\alias{plot_hmm_transitions}
\title{Transition Plots}
\usage{
plot_hmm_transitions(hmm_result)
}
\arguments{
\item{hmm_result}{run_hmm() returned object}
}
\value{
Transition diagram from final probabilities
}
\description{
provides a visual representation of trained hidden Markov model
}
\details{
Plots a HMM as a directed graph.
  States (rectangles) are interconnected by directed
  lines with line-weights proportional to the transition probabilities between
  the states.
}
\examples{
hmm_result = run_HMM(site_info = "mydata.sitelh",aln_info = "mydata.alninfo",model = 3)
plot_hmm_transitions(hmm_result)

}
