% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_mixtrees}
\alias{plot_mixtrees}
\title{Pediction Plots}
\usage{
plot_mixtrees(pred)
}
\arguments{
\item{pred}{predict_tree/predict_tree_mixed returned object}
}
\value{
input, viterbi and posterior decoding predicted path plot
}
\description{
Pediction Plots
}
\examples{
plot_mixtures(pred)
}
