% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_predictions}
\alias{plot_predictions}
\title{Prediction Plots}
\usage{
plot_predictions(hmm_result)
}
\arguments{
\item{hmm_result}{run_hmm() returned object}
}
\value{
hmm_input, viterbi or posterior decoding predicted path plot
}
\description{
Prediction Plots
}
\examples{
hmm_result = run_HMM(site_info = "mydata.sitelh",aln_info = "mydata.alninfo",model = 3)
plot_predictions(hmm_result)
}
