% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Initial Scatter Plots}
\usage{
plot_scatter(input_filepath, span = 0.03)
}
\arguments{
\item{input_filepath}{file path to a .sitelh or a .siteprob file from IQ-TREE}

\item{span}{the span of the kernel density plot (default 0.03)}
}
\value{
ggplot object with a scatter plot and smoothed kernel density
}
\description{
Before running the HMM, it's a really good idea to just look at the raw data from IQ-TREE,
to see for yourself if neighboring sites really do have similar class assignments
}
\examples{
plot_scatter(input_filepath = "sample_data.sitelh")
plot_scatter(input_filepath = "sample_data.siteprob")
}
