% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BW-models.R
\name{predict_class}
\alias{predict_class}
\title{Predict Class function}
\usage{
predict_class(sitein, alninfo, model, iter)
}
\arguments{
\item{sitein}{relative path of the input .sitelh/.siteprob file}

\item{alninfo}{relative path of the alignment file}

\item{model}{Choose model 1,2,3 or 4. Default=4}

\item{iter}{maximum iterations. Default=10000}
}
\value{
viterbi path, posterior path and convergence truth as a list
}
\description{
Predict Class function
}
\examples{
pred=predict_class("./iqtree/sample.siteprob","./iqtree/sample.alninfo",3)
v.pred<- pred[[1]];p.pred<-pred[[2]];conv<-pred[[3]]
}
