% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BW-models.R
\name{predict_class_mixed}
\alias{predict_class_mixed}
\title{Predict Class Mixed function}
\usage{
predict_class_mixed(sitein, alninfo, switch, iter)
}
\arguments{
\item{sitein}{relative path of the input .sitelh/.siteprob file}

\item{alninfo}{relative path of the alignment file}

\item{iter}{maximum iterations. Default=10000}
}
\value{
viterbi path, posterior path and convergence truth as a list
}
\description{
Predict Class Mixed function
}
\examples{
pred=predict_class("data100.2t.10k.fa.siteprob","data100.2t.10k.fa.siteprob",3)
v.pred<- pred[[1]];p.pred<-pred[[2]];conv<-pred[[3]]

}
