% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BW-models.R
\name{predict_tree}
\alias{predict_tree}
\title{Predict Tree function}
\usage{
predict_tree(file, model, variable, iter)
}
\arguments{
\item{file}{relative path of the input sitelh file}

\item{model}{Choose model 1,2,3 or 4. Default=4}

\item{variable}{post.prob.tree. or log.like.tree. Default= log.like.tree}

\item{iter}{maximum iterations. Default=10000}
}
\value{
viterbi path, posterior path and convergence truth as a list
}
\description{
Predict Tree function
}
\examples{
pred=predict_tree("../AliSim/taxa4/taxa4.data1.fa.sitelh",4,"post.prob.tree.")
v.pred<- pred[[1]];p.pred<-pred[[2]];conv<-pred[[3]]

}
