% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BW-mix_model.R
\name{predict_tree_mixed}
\alias{predict_tree_mixed}
\title{Predict Tree Mixed function}
\usage{
predict_tree_mixed(file, variable, iter)
}
\arguments{
\item{file}{relative path of the input sitelh file}

\item{variable}{post.prob.tree. or log.like.tree. Default= log.like.tree}

\item{iter}{maximum iterations. Default=10000}
}
\value{
viterbi path and posterior path as a list
}
\description{
Predict Tree Mixed function
}
\examples{
pred=predict_tree_mixed("../AliSim/taxa4/taxa4.data1.fa.sitelh","post.prob.tree.")
v.pred<- pred[[1]];p.pred<-pred[[2]]

}
