% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BW-models.R
\name{run_HMM}
\alias{run_HMM}
\title{Run_HMM to Post-processes the output of phylogenetic mixture models using HMMs}
\usage{
run_HMM(site_info, aln_info, model = 4, iter = 10000, algorithm = "viterbi")
}
\arguments{
\item{site_info}{relative path of the input .sitelh or .siteprob file from your IQ-TREE analysis.
In almost all of our tests and simulations, we found that the HMM performs better using site likelihoods(.sitelh file) rather than site posterior probabilities (the .siteprob file), so we recommend using it here.}

\item{aln_info}{relative path of the alignment file from IQ-TREE analysis.}

\item{model}{Choose model 1,2,3 or 4. Default=4}

\item{iter}{The maximum number of EM iterations to carry out before the cycling process is terminated and the partially trained model is returned. Defaults to 100.
The maximum change in log likelihood between EM iterations before the cycling procedure is terminated is set to 1E-07.}

\item{algorithm}{Select dynamic algorithm to predict class boundaries from the trained HMM model parameters.}
}
\value{
an object of class MixtureModelHMM
}
\description{
The function implements the Baum-Welch algorithm for training the HMM and
uses dynamic programming to to assign every site in the alignment to a class.
}
\details{
Model details

Model 1 has no additional emissions. Emissions = classes.

Model 2 has one additional emissions for constant sites. Emissions = classes + 1.

Model 3 has two additional emissions for constant sites and non-informative sites. Emissions = classes + 2.

Model 4 has three additional emissions for constant sites, non-informative sites and same parsimony sites. Emissions = classes + 3.

Detail information about approach https://github.com/roblanf/MixtureModelHMM
}
\examples{
hmm_result = run_HMM(site_info = "mydata.sitelh",aln_info = "mydata.alninfo")
}
