% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tables.R
\name{save_partition_file}
\alias{save_partition_file}
\title{Generate partition file}
\usage{
save_partition_file(hmm_result, output_filename)
}
\arguments{
\item{hmm_result}{run_hmm() returned object}

\item{output_filename}{name of file to be saved}
}
\value{
saves partition file in .nex format
}
\description{
Generate partition file
}
\examples{
hmm_result = run_HMM(site_info = "mydata.sitelh",aln_info = "mydata.alninfo",model = 3)
partition_file(hmm_result,"hmm_partitions")
}
