% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tables.R
\name{save_partitioning_scheme}
\alias{save_partitioning_scheme}
\title{Generate partition file}
\usage{
save_partitioning_scheme(hmm_result, output_filename)
}
\arguments{
\item{hmm_result}{run_hmm() returned object}

\item{output_filename}{name of file to be saved}
}
\value{
saves partition file
}
\description{
Generate partition file
}
\examples{
hmm_result = run_HMM(site_info = "mydata.sitelh",aln_info = "mydata.alninfo",model = 3)
save_partitioning_scheme(hmm_result,"hmm_partitions")
}
