% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tables.R
\name{summary.MixtureModelHMM}
\alias{summary.MixtureModelHMM}
\title{Summary}
\usage{
\method{summary}{MixtureModelHMM}(hmm_result, ...)
}
\arguments{
\item{hmm_result}{run_hmm() returned object}

\item{...}{additional arguments to be passed}
}
\value{
summary
}
\description{
Summary
}
\examples{
summary(hmm_result)
}
