% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tables.R
\name{transition_table}
\alias{transition_table}
\title{Create a transition table}
\usage{
transition_table(hmm_result)
}
\arguments{
\item{hmm_result}{run_hmm() returned object}
}
\value{
dataframe of transition table
}
\description{
Where each row has:
*site
*class_to
*class_from
And we ignore the first site, because it's not a transition.
}
\examples{
transition_table(hmm_result)
}
