% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_hmm_transitions}
\alias{plot_hmm_transitions}
\title{Transition Plots}
\usage{
plot_hmm_transitions(hmm_result)
}
\arguments{
\item{hmm_result}{run_hmm() returned object}
}
\value{
Transition diagram from final probabilities
}
\description{
Transition Plots
}
\examples{
plot_hmm_transitions(hmm_result)

}
