% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_predictions}
\alias{plot_predictions}
\title{Prediction Plots}
\usage{
plot_predictions(pred)
}
\arguments{
\item{pred}{predict_class/predict_class_mixed returned list}
}
\value{
input, viterbi and posterior decoding predicted path plot
}
\description{
Prediction Plots
}
\examples{
plot_predictions(pred)
}
