% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Inital Plots}
\usage{
plot_scatter(file, variable)
}
\arguments{
\item{file}{relative path of the input sitelh file}

\item{variable}{post.prob.tree. or log.like.tree. Default= post.prob.tree.}
}
\value{
scatter plot for initial post.prob.tree./log.like.tree.
}
\description{
Inital Plots
}
\examples{
plot_scatter("../AliSim/taxa4/taxa4.data1.fa.sitelh","post.prob.tree.")
}
