% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Inital Plots}
\usage{
plot_scatter(pred, variable)
}
\arguments{
\item{pred}{predict_tree/predict_tree_mixed returned object}

\item{variable}{post.prob.tree. or log.like.tree. Default= post.prob.tree.}
}
\value{
scatter plot for initial post.prob.tree./log.like.tree.
}
\description{
Inital Plots
}
\examples{
plot_scatter(pred,"post.prob.tree.")
}
