% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Inital Plots}
\usage{
plot_scatter(sitein)
}
\arguments{
\item{file}{relative path of the input sitelh/siteprob file}
}
\value{
scatter plot for initial log-likelihood/posterior probabilities
}
\description{
Inital Plots
}
\examples{
plot_scatter("sample_data.sitelh")
plot_scatter("sample_data.siteprob")
}
