% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_transitions}
\alias{plot_transitions}
\title{Transition Plots}
\usage{
plot_transitions(pred)
}
\arguments{
\item{pred}{predict_class/predict_class_mixed returned list}
}
\value{
Transition diagram from final probabilities
}
\description{
Transition Plots
}
\examples{
plot_transitions(pred)

}
