% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BW-models.R
\name{run_HMM}
\alias{run_HMM}
\title{Predict Class function}
\usage{
run_HMM(site_info, aln_info, model = 4, iter = 10000, algorithm = "viterbi")
}
\arguments{
\item{site_info}{relative path of the input .sitelh/.siteprob file}

\item{aln_info}{relative path of the alignment file}

\item{model}{Choose model 1,2,3 or 4. Default=4}

\item{iter}{maximum iterations. Default=10000}

\item{algorithm}{viterbi/posterior algoithm to be used to estimate final sequence of classes}
}
\value{
viterbi path or posterior path as classification as a list
}
\description{
Predict Class function
}
\examples{
hmm_result = run_HMM(site_info = "mydata.sitelh",aln_info = "mydata.alninfo",model = 3)
classification <- pred[[1]]
}
