% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_output.R
\name{save_file}
\alias{save_file}
\title{Saves the predicted output to gz file}
\usage{
save_file(hmm_result, output_filename)
}
\arguments{
\item{hmm_result}{run_hmm() returned object}

\item{output_filename}{name of file to be saved}
}
\value{
saves prediction output in .gz file
}
\description{
Saves the predicted output to gz file
}
\examples{
hmm_result = run_HMM(site_info = "mydata.sitelh",aln_info = "mydata.alninfo",model = 3)
save_file(hmm_result,"output")
classification<-system("gzcat output.gz",intern=TRUE)
}
