% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_output.R
\name{save_file}
\alias{save_file}
\title{Saves the predicted output to gz file}
\usage{
save_file(pred, algo, file_name)
}
\arguments{
\item{pred}{predict_tree/predict_tree_mixed returned list}

\item{algo}{"viterbi" or "posterior"}

\item{file_name}{name of file to be saved}
}
\value{
saves prediction output in .gz file
}
\description{
Saves the predicted output to gz file
}
\examples{
save_file(pred,"viterbi","output_v")
save_file(pred,"posterior","output_p")
v.pred<-system("gzcat output_v.gz",intern=TRUE)
}
