% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_output.R
\name{save_report}
\alias{save_report}
\title{Generates report}
\usage{
save_report(hmm_result, output_filename)
}
\arguments{
\item{hmm_result}{run_hmm() returned object}

\item{output_filename}{name of file to be saved}
}
\value{
saves prediction report in HTML file
}
\description{
Generates report
}
\examples{
hmm_result = run_HMM(site_info = "mydata.sitelh",aln_info = "mydata.alninfo",model = 3)
save_report(hmm_result,"hmm_report.Rmd")
}
