% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_output.R
\name{save_report}
\alias{save_report}
\title{Generates report}
\usage{
save_report(pred, filename)
}
\arguments{
\item{pred}{predict_tree/predict_tree_mixed returned list}

\item{file_name}{name of file to be saved}
}
\value{
saves prediction report in .txt
}
\description{
Generates report
}
\examples{
save_report(pred,"report")
}
